ディレクトリ構造変更、コメント追加
