+++ controller/notify/notify.go
@@ -1,14 +1,16 @@
-package controller
+package notify
 
 import (
+	"Gin_MVC/controller/login"
 	"log"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
 
+//GetNotify 通知取得
 func GetNotify(c *gin.Context) {
-	user, _, err := getLoginedUser(c)
+	user, _, err := login.GetLoginUser(c)
 	log.Println("notify", err)
 	if err != nil {
 		c.JSON(403, "")
